feat: share subagent preview card for task runs
